requirement.txt
mysql-connector-python
Pillow
opencv-python
imutils